Change from jre17 to jre11 in the image
Why:
The flink operator currently targets java11 based images by default and
if we provide a java17 based image then the operator or user has to
supply some add-opens parameters to the java command. To avoid having to
configure this we will target 11 for now.

The dream was to prepare for Flink 2.0+ where originally it was planned
to make java17 the default, but that appears to have dropped out of the
2.0 release. So using the same JRE as upstream is going to make our
lives easier.

Note we still build with JDK17 but target 11 as the release for the job
jar, so it's compatible.

Signed-off-by: Robert Young <robeyoun@redhat.com>